+++ src/kaltura-clients/index.ts
@@ -1,5 +1,3 @@
-export { KalturaHttpClientConfiguration } from './kaltura-http-client-configuration';
-export { KalturaClientBaseConfiguration} from './kaltura-client-base-configuration';
-export { KalturaBrowserHttpClient } from './kaltura-browser-http-client';
-export { KalturaClientBase } from './kaltura-client-base';
-export { KalturaHttpClientBase } from './kaltura-http-client-base';
+export * from './kaltura-browser-http-client';
+export * from './kaltura-client-base';
+export * from './kaltura-http-client-base';